Worked on Level 2
Implemented a character movement system which allows controlling the character in road left and right between the road bounds.
Created two different prefabs which are the city parts. Finished working on the appearance of one of them.
Implemented a system which will use the city parts as the character continues to run down the road. City parts will spawn according to player's position.